fixing agp bugs, refactoring
